Add files via upload
Car Price Prediction App
This project is a machine learning-based web application that predicts the selling price of a used car based on inputs like year of purchase, kilometers driven, fuel type, transmission type, and more. It uses a regression model trained on historical car data and is deployed using Streamlit for an interactive user experience.

Built with Python, scikit-learn, and Streamlit.